# apps/api-server/src/routes/general.rs
use crate::AppState;
use aiome_core::traits::JobQueue;
use axum::{
    extract::Path, extract::State, http::StatusCode, response::IntoResponse, response::Json,
    routing::get,
};
use shared::health::{HealthMonitor, ResourceStatus};
use std::fs;

pub async fn list_wiki_files(State(state): State<AppState>) -> Json<Vec<String>> {
    let mut files = Vec::new();
    if let Ok(entries) = fs::read_dir(&state.docs_path) {
        for entry in entries.flatten() {
            if let Some(name) = entry.file_name().to_str() {
                if name.ends_with(".md") {
                    files.push(name.to_string());
                }
            }
        }
    }
    files.sort();
    Json(files)
}

pub async fn get_wiki_content(
    State(state): State<AppState>,
    Path(filename): Path<String>,
) -> impl IntoResponse {
    if filename.contains("..") || filename.contains('/') || filename.contains('\\') {
        return (StatusCode::BAD_REQUEST, "Invalid filename").into_response();
    }

    let path = std::path::PathBuf::from(&state.docs_path).join(filename);
    match fs::read_to_string(path) {
        Ok(content) => content.into_response(),
        Err(_) => (StatusCode::NOT_FOUND, "Wiki not found").into_response(),
    }
}

#[utoipa::path(
    get,
    path = "/api/health",
    responses(
        (status = 200, description = "Get current system and agent health status", body = ResourceStatus),
        (status = 401, description = "Unauthorized")
    ),
    security(("api_key" = []))
)]
pub async fn get_health_status(
    State(state): State<AppState>,
    _auth: crate::auth::Authenticated,
) -> Json<ResourceStatus> {
    let mut monitor = state.health_monitor.lock().await;
    let mut status = monitor.check();

    // Fetch real agent stats
    if let Ok(stats) = state.job_queue.get_agent_stats().await {
        status.level = stats.level;
        status.exp = stats.exp;
        status.resonance = stats.resonance;
        status.creativity = stats.creativity;
        status.fatigue = stats.fatigue;
    }

    Json(status)
}

#[derive(serde::Serialize, sqlx::FromRow, utoipa::ToSchema)]
pub struct LogEntryResponse {
    pub id: i64,
    pub timestamp: Option<String>,
    pub level: String,
    pub target: String,
    pub message: String,
}

#[utoipa::path(
    get,
    path = "/api/v1/logs",
    params(
        ("limit" = Option<i64>, Query, description = "Maximum number of logs to return")
    ),
    responses(
        (status = 200, description = "Fetch application logs", body = Vec<LogEntryResponse>)
    )
)]
pub async fn get_logs(
    State(state): State<AppState>,
    _auth: crate::auth::Authenticated,
    axum::extract::Query(params): axum::extract::Query<std::collections::HashMap<String, String>>,
) -> impl IntoResponse {
    let limit = params
        .get("limit")
        .and_then(|s| s.parse::<i64>().ok())
        .unwrap_or(100);

    let pool = state.job_queue.get_pool();
    let rows = sqlx::query_as::<_, LogEntryResponse>(
        "SELECT id, timestamp, level, target, message FROM app_logs ORDER BY id DESC LIMIT ?",
    )
    .bind(limit)
    .fetch_all(pool)
    .await;

    match rows {
        Ok(logs) => (StatusCode::OK, Json(logs)).into_response(),
        Err(e) => (
            StatusCode::INTERNAL_SERVER_ERROR,
            format!("Failed to fetch logs: {}", e),
        )
            .into_response(),
    }
}