# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aiome Management Console API",
    "description": "Core API for the Autonomous AI Operating System",
    "license": {
      "name": "ELv2",
      "identifier": "ELv2"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/artifacts": {
      "get": {
        "tags": [
          "crate::routes::artifacts"
        ],
        "operationId": "list_artifacts_handler",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "description": "Filter by category",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "in": "query",
            "description": "Semantic search query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Limit results",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of artifacts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {}
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/artifacts/{id}": {
      "get": {
        "tags": [
          "crate::routes::artifacts"
        ],
        "operationId": "get_artifact_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Artifact ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Artifact metadata",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Artifact not found"
          }
        }
      },
      "delete": {
        "tags": [
          "crate::routes::artifacts"
        ],
        "operationId": "delete_artifact_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Artifact ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Artifact deleted"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
    "/api/v1/artifacts/{id}/edges": {
      "get": {
        "tags": [
          "crate::routes::artifacts"
        ],
        "operationId": "get_artifact_edges_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Artifact ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Artifact edges (graph)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {}
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/artifacts/{id}/files/{filename}": {
      "get": {
        "tags": [
          "crate::routes::artifacts"
        ],
        "operationId": "download_artifact_file_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Artifact ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filename",
            "in": "path",
            "description": "File name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "File content stream"
          },
          "400": {
            "description": "Invalid filename"
          },
          "404": {
            "description": "File not found"
          }
        }
      }
    },
...
    "/api/v1/expression/auto": {
      "post": {
        "tags": [
          "crate::routes::expression"
        ],
        "operationId": "toggle_auto_expression",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutoToggle"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Toggled auto-expression",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/v1/expression/generate": {
      "post": {
        "tags": [
          "crate::routes::expression"
        ],
        "operationId": "generate_expression",
        "responses": {
          "200": {
            "description": "Generated expression",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "No karma available"
          }
        }
      }
    },
    "/api/v1/expression/list": {
      "get": {
        "tags": [
          "crate::routes::expression"
        ],
        "operationId": "list_expressions",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Limit results",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Recent expressions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {}
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/logs": {
      "get": {
        "tags": [
          "crate::routes::general"
        ],
        "operationId": "get_logs",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of logs to return",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fetch application logs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LogEntryResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AutoToggle": {
        "type": "object",
        "required": [
          "enabled"
        ],
        "properties": {
          "enabled": {
            "type": "boolean"
          }
        }
      },
...
      "LogEntryResponse": {
        "type": "object",
        "required": [
          "id",
          "level",
          "target",
          "message"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "level": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "target": {
            "type": "string"
          },
          "timestamp": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API_SERVER_SECRET"
      }
    }
  }
}